third_party/WebKit/LayoutTests/external/wpt/WebIDL/ecmascript-binding/legacy-platform-object-expected.txt

This is a testharness.js-based test.
FAIL [[GetOwnProperty]] with getters and no settersassert_equals: (assert_prop_desc_equals: property 'enumerable') expected false but got true
PASS [[GetOwnProperty]] with named property getters and setters
PASS [[GetOwnProperty]] with indexed property getters and setters
PASS Test [[DefineOwnProperty]] with no indexed property setter support.
PASS Test [[DefineOwnProperty]] with indexed property setter support.
Harness: the test ran to completion.
